refactor: Tidy up mock classes
