+++ src/bot/play/bring/pr.test.ts
@@ -1,5 +1,5 @@
 import { MessageEmbed } from "discord.js";
-import { MockMessage } from "../../skin/mock-message";
+import { MockMessage } from "../../skin/mock/message";
 import { analectaForTest } from "../../skin/test-analecta";
 import { bringPR } from "./pr";
 import { colorFromState } from "../../exp/state-color";
