+++ src/bot/play/bring/repo.test.ts
@@ -1,5 +1,5 @@
 import { MessageEmbed } from "discord.js";
-import { MockMessage } from "../../skin/mock-message";
+import { MockMessage } from "../../skin/mock/message";
 import { analectaForTest } from "../../skin/test-analecta";
 import { bringRepo } from "./repo";
