+++ src/bot/play/flavor.test.ts
@@ -1,4 +1,4 @@
-import { MockMessage } from "../skin/mock-message";
+import { MockMessage } from "../skin/mock/message";
 import { analectaForTest } from "../skin/test-analecta";
 import { flavor } from "./flavor";
