+++ src/bot/play/subscribe/unsubscribe-notification.test.ts
@@ -1,6 +1,6 @@
-import { MockUserDB, placeholder } from "../../skin/mock-user-db";
+import { MockUserDB, placeholder } from "../../skin/mock/user-db";
 import { DiscordId } from "../../exp/discord-id";
-import { MockMessage } from "../../skin/mock-message";
+import { MockMessage } from "../../skin/mock/message";
 import { analectaForTest } from "../../skin/test-analecta";
 import { unsubNotification } from "./unsubscribe-notification";
